+++ src_client/platform_ios/iosPluginTexture.m
@@ -9,7 +9,7 @@
 // ----------------------------------------------------------------
 
 // ローカルデータ読み込み
-void platformPluginTextureLocal(void *param, char *src, void(*callback)(void *param, uint32_t glId, uint16_t texw, uint16_t texh, uint16_t imgw, uint16_t imgh)){
+void platformPluginTextureLocal(void *param, char *src, void(*callback)(void *param, int glId, int texw, int texh, int imgw, int imgh)){
 	pluginTextureCallbackId callbackId = gamePluginTextureCallbackSet(param, callback);
 	NSString *nsstr1 = [NSString stringWithCString: src encoding: NSUTF8StringEncoding];
 	[IosPluginTexture platformPluginTextureLocal: callbackId src: nsstr1];
